flowableUI代码优化
(cherry picked from commit 00c1013aca1ecd38294aa650f2e2695cad8d45ce)

+++ y9-module-flowableui/risenet-y9boot-webapp-flowableui-position/src/main/java/net/risesoft/util/DeletePrintPdfFileUtil.java
@@ -1,17 +1,16 @@
 package net.risesoft.util;
 
-import java.io.File;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Date;
-
+import lombok.extern.slf4j.Slf4j;
+import net.risesoft.y9.Y9Context;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 
-import lombok.extern.slf4j.Slf4j;
 
-import net.risesoft.y9.Y9Context;
+import java.io.File;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Date;
 
 @Configuration
 @EnableScheduling
@@ -44,17 +43,19 @@ public class DeletePrintPdfFileUtil {
             File realFile = new File(pdfPath);
             if (realFile.exists() && realFile.isDirectory()) {
                 File[] subfiles = realFile.listFiles();
-                for (File file : subfiles) {
-                    if (file.getName().contains(".pdf")) {
-                        long lastModified = file.lastModified();// pdf最后更新时间
-                        if (lastModified < timeInMillis) {// 删除当前时间前5个小时的pdf文件
-                            file.delete();
+                if (subfiles != null) {
+                    for (File file : subfiles) {
+                        if (file.getName().contains(".pdf")) {
+                            long lastModified = file.lastModified();// pdf最后更新时间
+                            if (lastModified < timeInMillis) {// 删除当前时间前5个小时的pdf文件
+                                file.delete();
+                            }
                         }
                     }
                 }
             }
         } catch (Exception e) {
-            e.printStackTrace();
+            LOGGER.error("定时任务删除pdf文件异常：{}", e.getMessage());
         }
     }
